# Skip long string partitioning in string argsort