--- a/test/StringArgsortSpeedTest.chpl
+++ b/test/StringArgsortSpeedTest.chpl
@@ -2,7 +2,7 @@ use TestBase;
 
 config const size = 10**4;
 config const minLen = 1;
-config const maxLen = 8;
+config const maxLen = 16;
 
 proc main() {
   var st = new owned SymTab();